Separate AWS clients into client and s3doc
