+++ app/api/api_v1/routers/search.py
@@ -18,7 +18,8 @@ from sqlalchemy.orm import Session
 from starlette.responses import RedirectResponse
 
 from app.api.api_v1.schemas.search import SearchRequestBody, SearchResponse
-from app.clients.aws.aws import S3Client, S3Document, get_s3_client
+from app.clients.aws.client import S3Client, get_s3_client
+from app.clients.aws.s3 import S3Document
 from app.clients.db.session import get_db
 from app.config import (
     AWS_REGION,
