+++ tests/conftest.py
@@ -15,7 +15,7 @@ from sqlalchemy.engine import Engine
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy_utils import create_database, database_exists, drop_database
 
-from app.clients.aws.aws import S3Client, get_s3_client
+from app.clients.aws.client import S3Client, get_s3_client
 from app.clients.db.session import get_db
 from app.core import custom_app, security
 from app.core.custom_app import AppTokenFactory
